fix SignalDispatcher::terminate, we should try until `$wait` seconds later if we can not terminate process, and return the final results.

+++ src/Commands/StopCommand.php
@@ -58,7 +58,11 @@ class StopCommand extends Command
 
         $this->info('Stopping server...');
 
-        $inspector->stopServer();
+        if (!$inspector->stopServer()) {
+            $this->error('Stopping server fails.');
+
+            return 1;
+        }
 
         app(SwooleServerStateFile::class)->delete();
 
@@ -84,7 +88,11 @@ class StopCommand extends Command
 
         $this->info('Stopping server...');
 
-        $inspector->stopServer();
+        if (!$inspector->stopServer()) {
+            $this->error('Stopping server fails.');
+
+            return 1;
+        }
 
         app(RoadRunnerServerStateFile::class)->delete();
